+++ api-spec/src/main/kotlin/eu/vendeli/rethis/api/spec/commands/server/ShutdownCommand.kt
@@ -3,7 +3,6 @@ package eu.vendeli.rethis.api.spec.commands.server
 import eu.vendeli.rethis.shared.annotations.RIgnoreSpecAbsence
 import eu.vendeli.rethis.shared.annotations.RedisCommand
 import eu.vendeli.rethis.shared.annotations.RedisMeta
-import eu.vendeli.rethis.shared.decoders.geospatial.GeoPosDecoder
 import eu.vendeli.rethis.shared.decoders.server.ShutdownDecoder
 import eu.vendeli.rethis.shared.request.server.SaveSelector
 import eu.vendeli.rethis.shared.request.server.ShutdownOptions