+++ client/src/commonMain/kotlin/eu/vendeli/rethis/command/hash/Hget.kt
@@ -1,18 +1,17 @@
 package eu.vendeli.rethis.command.hash
 
 import eu.vendeli.rethis.ReThis
-import eu.vendeli.rethis.codecs.hash.HGetBSCommandCodec
+import eu.vendeli.rethis.codecs.hash.HGetBACommandCodec
 import eu.vendeli.rethis.codecs.hash.HGetCommandCodec
 import eu.vendeli.rethis.topology.handle
-import kotlinx.io.bytestring.ByteString
 
-public suspend fun ReThis.hGetBS(key: String, `field`: String): ByteString? {
+public suspend fun ReThis.hGetBA(key: String, `field`: String): ByteArray? {
     val request = if(cfg.withSlots) {
-        HGetBSCommandCodec.encodeWithSlot(charset = cfg.charset, key = key, field = field)
+        HGetBACommandCodec.encodeWithSlot(charset = cfg.charset, key = key, field = field)
     } else {
-        HGetBSCommandCodec.encode(charset = cfg.charset, key = key, field = field)
+        HGetBACommandCodec.encode(charset = cfg.charset, key = key, field = field)
     }
-    return HGetBSCommandCodec.decode(topology.handle(request), cfg.charset)
+    return HGetBACommandCodec.decode(topology.handle(request), cfg.charset)
 }
 
 public suspend fun ReThis.hGet(key: String, `field`: String): String? {