+++ client/src/commonMain/kotlin/eu/vendeli/rethis/command/list/Blmove.kt
@@ -1,11 +1,10 @@
 package eu.vendeli.rethis.command.list
 
 import eu.vendeli.rethis.ReThis
-import eu.vendeli.rethis.codecs.list.BlMoveBSCommandCodec
+import eu.vendeli.rethis.codecs.list.BlMoveBACommandCodec
 import eu.vendeli.rethis.codecs.list.BlMoveCommandCodec
 import eu.vendeli.rethis.shared.response.common.MoveDirection
 import eu.vendeli.rethis.topology.handle
-import kotlinx.io.bytestring.ByteString
 
 public suspend fun ReThis.blMove(
     source: String,
@@ -22,17 +21,17 @@ public suspend fun ReThis.blMove(
     return BlMoveCommandCodec.decode(topology.handle(request), cfg.charset)
 }
 
-public suspend fun ReThis.blMoveBS(
+public suspend fun ReThis.blMoveBA(
     source: String,
     destination: String,
     whereFrom: MoveDirection,
     whereTo: MoveDirection,
     timeout: Double,
-): ByteString? {
+): ByteArray? {
     val request = if(cfg.withSlots) {
-        BlMoveBSCommandCodec.encodeWithSlot(charset = cfg.charset, source = source, destination = destination, whereFrom = whereFrom, whereTo = whereTo, timeout = timeout)
+        BlMoveBACommandCodec.encodeWithSlot(charset = cfg.charset, source = source, destination = destination, whereFrom = whereFrom, whereTo = whereTo, timeout = timeout)
     } else {
-        BlMoveBSCommandCodec.encode(charset = cfg.charset, source = source, destination = destination, whereFrom = whereFrom, whereTo = whereTo, timeout = timeout)
+        BlMoveBACommandCodec.encode(charset = cfg.charset, source = source, destination = destination, whereFrom = whereFrom, whereTo = whereTo, timeout = timeout)
     }
-    return BlMoveBSCommandCodec.decode(topology.handle(request), cfg.charset)
+    return BlMoveBACommandCodec.decode(topology.handle(request), cfg.charset)
 }